pyproject.toml
[tool.poetry]
name = "model-orchestrator"
version = "0.1.0"
description = ""
authors = ["boutrig abdennour <abdennour.boutrig@crunchdao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
boto3 = "^1.35.90"
python-box = "^7.3.0"
attrs = "^24.3.0"
pyyaml = "^6.0.2"
google = "^3.0.0"
colorama = "^0.4.6"
requests = "^2.32.3"
deepdiff = "^8.1.1"
websockets = "^14.2"
aio-pika = "^9.5.4"
cryptography = "^44.0.1"
sqlite-utils = "^3.38"
newrelic = "^10.6.0"
pydantic = "^2.11.7"
docker = "^7.1.0"
click = "^8.2.1"
crunch-cli = "^7.1.0"
ruamel-yaml = "^0.18.14"
exceptiongroup = "^1.3.0"
opening-hours-py = "^1.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
deepdiff = "^8.2.0"
pytest-env = "^1.1.5"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
model-orchestrator = "model_orchestrator.cli:cli"
